+++ src/components/Profile/LanguageChooser.jsx
@@ -2,22 +2,28 @@ import React from 'react'
 import {Select, MenuItem, Typography} from "@material-ui/core"
 import {withTranslation} from "react-i18next"
 import {languages} from "../../locales/locales.json"
+import {useNotification} from '../../hooks'
 
 const LanguageChooser = ({i18n}) => {
+
+  const {notify} = useNotification()
 
   const changeLanguage = ({target: {value}}) => {
     localStorage.setItem("preferredLanguage", value)
     i18n.changeLanguage(value)
+    notify({name: "language"})
   }
 
   return(
     <Select
+      color="secondary"
       onChange={changeLanguage}
       value={i18n.language}
     >
       {languages.map(({code, title, flag}) =>
         <MenuItem
           key={code}
+          selected
           value={code}
         >
           <Typography variant="subtitle1">
@@ -29,4 +35,4 @@ const LanguageChooser = ({i18n}) => {
   )
 }
 
-export default withTranslation("settings")(LanguageChooser)
+export default withTranslation("profile")(LanguageChooser)