+++ src/components/index.js
@@ -1,0 +1,19 @@
+import OfflineStatus from "./OfflineStatus"
+import Dashboard from "./Dashboard"
+import Register from "./Register"
+import Landing from "./Landing"
+import NotFound from "./NotFound"
+import Profile from "./Profile"
+import Centered from "./Centered"
+
+export {Dialog, Notification} from "./shared"
+
+export {
+  Centered,
+  OfflineStatus,
+  Dashboard,
+  Register,
+  Landing,
+  NotFound,
+  Profile
+}