+++ src/components/shared/__tests__/DestructButton.test.js
@@ -1,0 +1,20 @@
+import {DestructButton} from ".."
+
+describe("DestructButton component", () => {
+  const props = {}
+  const wrapper = shallow(<DestructButton {...props}/>)
+
+  it("renders correctly", () => {
+    expect(wrapper.dive()).toHaveLength(1)
+  })
+
+  it("children propagated", () => {
+    wrapper.setProps({children: "test"})
+    expect(wrapper.dive().prop("children")).toBe("test")
+  })
+
+  it("Button props propagated", () => {
+    wrapper.setProps({color: "primary"})
+    expect(wrapper.dive().prop("color")).toBe("primary")
+  })
+})