+++ src/lib/i18next.js
@@ -3,40 +3,43 @@ import common_en from "../locales/en/common.json"
 import common_no from "../locales/no/common.json"
 import profile_en from "../locales/en/profile.json"
 import profile_no from "../locales/no/profile.json"
+import boats_en from "../locales/en/boats.json"
+import boats_no from "../locales/no/boats.json"
+import dashboard_en from "../locales/en/dashboard.json"
+import dashboard_no from "../locales/no/dashboard.json"
 import pages_en from "../locales/en/pages.json"
 import pages_no from "../locales/no/pages.json"
-import settings_en from "../locales/en/settings.json"
-import settings_no from "../locales/no/settings.json"
+import register_en from "../locales/en/register.json"
+import register_no from "../locales/no/register.json"
 import languages from "../locales/locales.json"
 
+const en = {
+  common: common_en,
+  profile: profile_en,
+  dashboard: dashboard_en,
+  boats: boats_en,
+  pages: pages_en,
+  register: register_en
+}
+
+const no = {
+  common: common_no,
+  profile: profile_no,
+  dashboard: dashboard_no,
+  boats: boats_no,
+  pages: pages_no,
+  register: register_no
+}
 
 i18next.init({
   lng: localStorage.getItem("preferredLanguage") || navigator.language || languages.default,
   resources: {
-    en: {
-      common: common_en,
-      profile: profile_en,
-      pages: pages_en,
-      settings: settings_en
-    },
-    no: {
-      common: common_no,
-      profile: profile_no,
-      pages: pages_no,
-      settings: settings_no
-    },
-    "ny-NO":{
-      common: common_no,
-      pages: pages_no,
-      profile: profile_no,
-      settings: settings_no
-    },
-    "nb-NO": {
-      common: common_no,
-      pages: pages_no,
-      profile: profile_no,
-      settings: settings_no
-    }
+    en,
+    "en-US": en,
+    "en-GB": en,
+    no,
+    "ny-NO": no,
+    "nb-NO": no
   }
 })