+++ CHANGELOG.md
@@ -4,9 +4,18 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased] 
 ### Added
 - User now gets a notification that the PWA can be added to homescreen
+- in-app notification system ([#40](https://github.com/ecatch-kyst/web/issues/40))
+
+
+## [1.0.0-beta1] - 2019-02-11
+### Added
+- bottom navigation
 - use `localStorage` to save preferred language over sessions
-- settings page
+- unified settings & profile page
 - dark mode
+    - use `localStorage` to save preferred dark/light mode over sessions
+- Dialog
+- DestructButton
 - authentication:
     - registration
     - login
@@ -17,8 +26,14 @@ All notable changes to this project will be documented in this file.
             - after registering a user -> dashboard
             - dashboard -> root
             - 404 -> root
+### Changed
+- Offline status
+- UI element sizes increased
 ### Removed
 - example codes
+### Fixed
+- don't show Navigation on Register
+- when the login form is submitted, the page does not refresh anymore
 
 ## [0.3.0] - 2019-01-30
 ### Added